+++ web/src/components/layout/Layout.jsx
@@ -23,7 +23,7 @@ export const Layout = ({ status, children }) => {
 export const LayoutUserDashboard = ({ children }) => {
     return (
         <>
-            <Container maxWidth="sm">{children}</Container>
+            <Container maxWidth="md">{children}</Container>
         </>
     );
 };
